Avoid DNS host checking with extended trust manager
Sun's SSL implementation now checks if the trust manager on the SSL
context is a ExtendedTrustManager, and if not it wraps it an an
AbstractTrustManager that does extra checking to ensure that the DNS
host on the request matches the specified hosts on the certificate.

This changes skuber insecure trust manager to be an extended trust
manager to avoid that extra DNS check. Otherwise, in test circumstances
where a certificate does not list the correct DNS that is used for the
host then we are unable to connect to the kubernetes cluster.