Make K8SException less vulnerable to annoying nulls.
The previous code could periodically end up in a state where we got a
"null" message when a validation error occured in Kubernetes. This new
implementation now surfaces *all* the relevant details of the failure in
the message field, which is more compatible with how standard logging
implementations will handle these exceptions when they are provided as
an argument.