Use collect instead of filter
Since Akka 2.6.2 `filter` will prefetch one element. This will not change overall output of a stream but changes timing of when demand is propagated. In `WatchSourceSpec` it will lead to an extra request being made even if there's no demand yet from the consumer. In general, the test should be hardened not to fail if these details change but for now this simple change will prevent the prefetching because `collect` does not (yet?) prefetch elements.

Great work, @seglo, pinpointing the problem to the filter change in https://github.com/akka/akka/pull/28467.