Use request-based Akka HTTP API to share connection pools
LongPollingPool was calling newHostConnectionPool for each materialization of
the 'watch'. While that is not wrong per se (pools are cleaned up when they
have been idle for some time), it can lead to a large number of pools in memory
when the Watch streams are regularly restarted.

By using the request-based API here, Akka HTTP will share connection pools
among requests with the same ConnectionContext, which should lead to less
resource usage in such scenario's.