Mread upgrade to akka 2.6.x (#314)
* upgrade sbt to 1.3.13

* add type Format[skuber.Protocol.Value] to line 405

* Upgrade Bouncy Castle to 1.66 and Play JSON to 2.9.0 due to security concerns

* Upgrade Akka (2.6.8) & Akka HTTP (10.1.12)

* inital refactoring - mostly removing implicit materializers

* finish refactoring - mostly removing implicit materializers

* drop support for Scala 2.11

* drop reference to Scala 2.11 for current release

* drop reference to Scala 2.11 for v2

* change import for MockitoSugar (org.scalatestplus.mockito.MockitoSugar)

* Use collect instead of filter

Since Akka 2.6.2 `filter` will prefetch one element. This will not change overall output of a stream but changes timing of when demand is propagated. In `WatchSourceSpec` it will lead to an extra request being made even if there's no demand yet from the consumer. In general, the test should be hardened not to fail if these details change but for now this simple change will prevent the prefetching because `collect` does not (yet?) prefetch elements.

Great work, @seglo, pinpointing the problem to the filter change in https://github.com/akka/akka/pull/28467.

* upgrade Mockito (v3.4.4) & Akka HTTP to (10.1.12)

* bump commons-io 2.7

Co-authored-by: Johannes Rudolph <johannes.rudolph@gmail.com>