feat: reload access tokens for in cluster config
Fixes #350

This implements token reloading for in cluster config. It's built on #424 so
that needs to be reviewed/merged first.

Also fixed a resource leak where the input streams for reading the files for
in cluster config were not being closed. This was necessary to fix given that
we're now frequently loading those streams.