Fixed error on missing config references (#425)
It is possible for kube config file to contain a context that refers to cluster
or user configurations that don't exist. I don't know how, but my kubeconfig
file does, and even when I clean it up, this frequently happens. These contexts
are old contexts that I don't use anymore, so they are safe to ignore. kubectl
just ignores them. Skuber should do the same.